Move to Formatter/Domain the PathFilter colaborator for the FormatCommand
