+++ src/php/Formatter/Domain/PathFilterInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Phel\Formatter\Command;
+namespace Phel\Formatter\Domain;
 
 interface PathFilterInterface
 {
