--- src/php/Formatter/Exceptions/CanNotRemoveAtTheTopException.php
@@ -1,11 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Phel\Formatter\Exceptions;
-
-use RuntimeException;
-
-final class CanNotRemoveAtTheTopException extends RuntimeException
-{
-}
