+++ src/php/Formatter/FormatterFactory.php
@@ -8,8 +8,8 @@ use Gacela\Framework\AbstractFactory;
 use Phel\Command\CommandFacadeInterface;
 use Phel\Compiler\CompilerFacade;
 use Phel\Formatter\Command\FormatCommand;
-use Phel\Formatter\Command\PathFilterInterface;
-use Phel\Formatter\Command\PhelPathFilter;
+use Phel\Formatter\Domain\PathFilterInterface;
+use Phel\Formatter\Domain\PhelPathFilter;
 use Phel\Formatter\Formatter\Formatter;
 use Phel\Formatter\Formatter\FormatterInterface;
 use Phel\Formatter\Rules\Indenter\BlockIndenter;
